Move field struct to its own file
